Buy2CashStoreMangBackend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.1
supabase==2.4.5
pydantic==2.5.0
python-dotenv==1.0.0
openai==1.12.0
redis==5.0.1
